# Improve Guid equality checks on 64-bit platforms