@BruceForstall 2020-04-30
Additional benchmarks: https://github.com/dotnet/performance/pull/1302