@EgorBo 2020-04-30
Is it safe for arm32 as well (accessing misaligned long)?