@EgorBo 2020-04-30
> Did you consider a [vectorized approach](https://sharplab.io/#v2:EYLgxg9gTgpgtADwGwBYA0AXEBDAzgWwB8ABABgAJiBGAOgCUBXAOwwEt8YaBhCfAB1YAbGFADKIgG6swMXAG4AsACgylWoxbtOASRZRWTXNPnLV1eszYcaujPsPGaADQAcSRUtMBmcrjsMwDHIAcQZWABNlAG9lcjjyPn0JbAwYcgMggH1sOXiAejzyABEIciYIINgmbA5yAApgA2woAE9fEVZsQVYALxTWCCYASlj4xNZk1N8AC2gs4FyC4tLyyphq2oam1vb9Lt7+wZGleISklLTcWagssEXCkrKK8iqatK3qndwO/b62I9GcXGkzSwBaU0y4Vy5CWj1WL3Wb3qjU+bW+e26fwGw0BZwmF3IYIhMGhsJWz1emxRzTRP0xhxxJzG5ymRLSmQAZqSHuS1ht3tSvnSDv9GadgQS2eRMgBzbnLJ58pEfGm7Tr00XHcUs0Hg9nTeVwimIqnbWkYkXYrXM/GsvXS1iG3kI/nIs1q34M61AnWE+2ZABWTsVLuVgvN6stAKZPttuohAGtg/DKQL3ejI1jo95CRAIIJyABRACODC6uDqoQi5BlxxiMZhhQAchAAO7kSD8ZppGDYMDTcgQDnkDDTHvCDgsXC44gAdnIAFVDNgOZwAIK4AA8GTQ5EEgxlAD46rBhzKaNkhnEALzXxfL1c0Dfbli7/dMI8nmDDy+4gBkf6nPeuAruuW5sm+B7HqeNYXuEV63sBoFPuBeqQR+0HftK8EeAAvsoObAHmBYlmWgi4FQlZhOENZ1riABqMCBNAVAAEwuJubKHuQEgUHeS4gY+z47uQjHMVAbEcVxmFnhe2BDB4pxiRgLHsZxercRIVDkPxD5gS+GC7spqlSRpMnSvJHgMUxKkSWpXHkDAxY6eQojfKx3C8Hw3akV0dS8buWkKbiGTkPgeAJi5bkwB5ACyEASDAsURXUTnBZ4DZzmFEU6XepAIAAYkVBV4couFAA==)?
> Only shown for SSE2, needs some specialization for Arm, and I don't know if it's worth it...

Last time I tried it was slower,
I think in most cases guids are different and early out check is enough.