@jkotas 2020-04-30
> Is it safe for arm32 as well (accessing misaligned long)?

+1. It would be better to use `Unsafe.ReadUnaligned` here to make the code portable.