@BruceForstall 2020-10-05
> @BruceForstall, are you still working on this? Thanks.

I haven't, and likely won't get back to it. I'll go ahead and close the PR. I'd be happy if someone picked it up. The "only thing" is to try Jan's suggestion about using `Unsafe.ReadUnaligned`, and verifying performance on all platforms still improves.